refactor: move code from query service to newly created builder service
